Prefix
Use prefix to record number needed to split 0 and 1. We could notice
that if we can delete a subsequence (delete it not longer because there
may have repeate 1 or 0 at the beginning or ending), it won't affect the
repeated 1 and 0 not in the subsequence. Actually, the number to delete
all is the maximum number of split 0 and split 1.